Revert code to the last successful deployment state and fix errors
Reverts all code files to the last known good deployment state and addresses build/deployment issues.
