--- AUTOSCALE_DEPLOYMENT_SOLUTION.md
@@ -1,33 +1,0 @@
-# Autoscale 배포 문제 완전 해결 가이드
-
-## 문제 진단
-1. **MODULE_NOT_FOUND 오류**: 번들링된 서버에서 동적 import 실패
-2. **Connection Refused 오류**: 서버가 127.0.0.1에 바인딩되어 프록시 접근 불가
-3. **비동기 초기화 블로킹**: 백그라운드 서비스가 서버 시작을 지연
-
-## 완전 해결 방법
-
-### 1. 서버 초기화 안정화 ✅
-- 동적 import 실패 시 경고만 표시하고 서버 시작 계속
-- 백그라운드 서비스를 비블로킹으로 초기화
-- 모든 모듈 로딩 오류에 대한 graceful handling
-
-### 2. Autoscale 호환 빌드 ✅
-- packages=external 옵션으로 외부 의존성 처리
-- NODE_ENV=production 빌드타임 설정
-- 완전한 package.json 의존성 포함
-
-### 3. 네트워크 바인딩 수정 ✅
-- 0.0.0.0:5000에 명시적 바인딩
-- HOST/PORT 환경변수 지원
-- 프로덕션 환경 포트 고정
-
-## 테스트 결과
-✅ 서버가 0.0.0.0:3001에서 정상 시작
-✅ MODULE_NOT_FOUND 오류 해결
-✅ 백그라운드 서비스 정상 초기화
-✅ 81.1kb 최적화된 번들 생성
-
-## 배포 명령어
-권장 빌드: `node autoscale-build.js`
-대안 빌드: `node production-build.js`
