--- CHANNEL_DELETION_DOCUMENTATION.md
@@ -1,55 +1,0 @@
-# 채널 삭제 및 영상 정리 로직 구현
-
-## 구현된 기능
-
-### 1. 완전한 채널 삭제 로직
-- **연관 데이터 순차 삭제**: 외래 키 제약조건을 준수하여 안전한 삭제
-- **삭제 순서**: 영상 업보트 → 영상 → 채널 구독 → 채널
-- **삭제 통계**: 삭제된 영상 수와 함께 상세 로그 출력
-
-### 2. 자동 캐시 정리 시스템
-- **영상별 캐시 관리**: 개별 영상의 처리 상태 추적
-- **채널 삭제 시 일괄 정리**: 관련 영상들의 캐시 자동 삭제
-- **재등록 시 깨끗한 시작**: 삭제된 영상이 다시 등록될 때 기존 캐시 영향 없음
-
-### 3. 데이터베이스 무결성 시스템
-- **시작 시 자동 검사**: 서버 시작 시 고아 레코드 자동 감지 및 정리
-- **일일 무결성 검사**: 매일 오전 3시 자동 데이터베이스 무결성 검사
-- **고아 레코드 처리**: 존재하지 않는 채널을 참조하는 영상/업보트/구독 자동 삭제
-
-### 4. RSS 수집기 개선
-- **중복 방지**: 동일한 영상이 다른 채널에 속하는 경우 감지 및 로깅
-- **삭제된 채널 처리**: YouTube에서 삭제된 채널 자동 감지 및 정리
-- **안전한 재처리**: 캐시 정리로 영상 재등록 시 올바른 처리 보장
-
-## 검증된 결과
-
-### 발견된 문제
-- **고아 레코드 발견**: 채널 0개, 영상 13개가 존재하지 않는 채널 ID 참조
-- **원인**: 이전 채널 삭제 시 관련 영상이 완전히 정리되지 않음
-
-### 해결 완료
-- **고아 영상 13개 삭제**: 존재하지 않는 채널 ID (49, 55, 56, 57) 참조 영상 정리
-- **현재 상태**: 채널 0개, 영상 0개로 데이터 일관성 확보
-- **자동 무결성 시스템**: 향후 고아 레코드 발생 방지
-
-### 로그 출력 예시
-```
-DELETE 13 (고아 영상 정리)
-Completely deleted channel 66 and all related data (8 videos removed, cache cleared)
-일일 데이터베이스 무결성 검사 시작...
-```
-
-## 사용자 인터페이스 개선
-
-### 채널 삭제 API
-- **마지막 구독자 확인**: 구독자가 1명 이하일 때만 완전 삭제
-- **점진적 삭제**: 구독 해제 → 완전 삭제 단계별 처리
-- **상세 응답**: 삭제 결과에 따른 적절한 메시지 반환
-
-### 에러 처리 강화
-- **외래 키 제약조건 준수**: 안전한 삭제 순서로 오류 방지
-- **자동 복구 메커니즘**: 서버 시작 시 데이터 무결성 자동 복구
-- **예방적 유지보수**: 일일 자동 검사로 문제 사전 방지
-
-이 구현으로 채널이 삭제될 때 모든 관련 데이터가 깨끗하게 정리되며, 고아 레코드가 발생해도 자동으로 감지하여 정리합니다.
