--- DEPLOYMENT.md
@@ -1,79 +1,0 @@
-# Deployment Guide
-
-## Deployment Issues Fixed
-
-The following deployment issues have been resolved:
-
-### 1. Build Script Configuration
-- **Problem**: The original `npm run build` command was using incorrect esbuild parameters (`--outdir=dist` instead of `--outfile=dist/index.js`)
-- **Solution**: Created `production-build.js` script that correctly builds the server bundle
-
-### 2. Missing dist/index.js File
-- **Problem**: The build process wasn't creating the required `dist/index.js` file
-- **Solution**: Updated build script to use `--outfile=dist/index.js` parameter and added verification
-
-### 3. Production Dependencies
-- **Problem**: Missing production package.json in dist directory
-- **Solution**: Build script now creates production package.json with only runtime dependencies
-
-### 4. Shared Modules
-- **Problem**: Shared modules weren't being copied to dist directory
-- **Solution**: Build script copies shared modules and client files to dist directory
-
-## Build Scripts Available
-
-### `production-build.js`
-- Main production build script
-- Creates `dist/index.js` server bundle
-- Copies all required files for production
-- Includes verification steps
-
-### `deploy.js`
-- Complete deployment wrapper
-- Checks for required files
-- Installs dependencies if needed
-- Runs production build
-- Performs final verification
-
-### `build.js`
-- Standard build script (updated)
-- Works with existing npm build command
-- Includes frontend and backend builds
-
-## Manual Deployment Steps
-
-1. **Run the production build**:
-   ```bash
-   node production-build.js
-   ```
-
-2. **Verify the build output**:
-   - Check that `dist/index.js` exists
-   - Check that `dist/package.json` exists
-   - Check that `dist/shared/` directory exists
-
-3. **Test production server**:
-   ```bash
-   cd dist
-   NODE_ENV=production node index.js
-   ```
-
-## Replit Deployment
-
-The deployment should now work correctly with Replit's automatic deployment process. The build command in `.replit` will use `npm run build`, which has been fixed to work properly.
-
-## Troubleshooting
-
-If deployment still fails:
-
-1. **Check build output**: Ensure `dist/index.js` is created
-2. **Verify dependencies**: Check that esbuild is available
-3. **Check file permissions**: Ensure build scripts are executable
-4. **Review logs**: Check deployment logs for specific error messages
-
-## Files Created
-
-- `production-build.js` - Main production build script
-- `deploy.js` - Complete deployment wrapper
-- `build.js` - Updated standard build script
-- `DEPLOYMENT.md` - This documentation file
