--- DEPLOYMENT_CONNECTION_FIX.md
@@ -1,66 +1,0 @@
-# 배포 연결 거부 오류 수정 가이드
-
-## 문제 분석
-`error proxying request error=dial tcp 127.0.0.1:5000: connect: connection refused` 오류는 다음과 같은 원인으로 발생합니다:
-
-1. **서버가 127.0.0.1에만 바인딩**: 배포 환경에서 프록시가 접근할 수 없음
-2. **비동기 초기화 블로킹**: 데이터베이스/RSS 초기화가 서버 시작을 지연시킴
-3. **포트 충돌 처리**: 배포 환경에서 포트 변경 불가
-4. **환경 변수 미설정**: NODE_ENV=production 미적용
-
-## 적용된 수정사항
-
-### 1. 서버 바인딩 개선 ✅
-- 모든 인터페이스(0.0.0.0)에 바인딩하도록 명시적 설정
-- HOST 환경변수 지원 추가
-- 배포 환경에서 적절한 호스트 설정 보장
-
-### 2. 비동기 초기화 최적화 ✅
-- 서버 시작을 블로킹하는 데이터베이스 초기화를 백그라운드로 이동
-- RSS 수집기 초기화 지연
-- 서버 시작 후 백그라운드 서비스 실행
-
-### 3. 배포별 포트 처리 ✅
-- 프로덕션 환경에서 포트 변경 방지
-- 포트 충돌 시 명확한 오류 메시지
-- PORT 환경변수 우선 사용
-
-### 4. 향상된 로깅 ✅
-- 서버 시작 과정의 상세한 로깅
-- 환경 설정 확인
-- 연결 상태 디버깅 정보
-
-## 수정된 파일들
-
-### server/index.ts
-- 0.0.0.0 호스트 바인딩 명시
-- 백그라운드 서비스 초기화
-- 배포 환경별 오류 처리
-- 향상된 로깅 시스템
-
-### production-build.js
-- NODE_ENV=production 빌드타임 설정
-- 최적화된 빌드 스크립트
-- 배포용 package.json 생성
-
-### deployment-fix.js
-- 배포 전용 빌드 스크립트
-- 연결 문제 해결에 특화
-- 시작 스크립트 생성
-
-## 검증 결과
-
-✅ 서버가 0.0.0.0:5000에 올바르게 바인딩
-✅ 프로덕션 환경 변수 설정 확인
-✅ 빌드 스크립트 정상 작동 (80.7kb)
-✅ 백그라운드 서비스 비블로킹 초기화
-✅ 향상된 오류 처리 및 로깅
-
-## 배포 권장사항
-
-1. **빌드 명령어**: `node production-build.js`
-2. **시작 명령어**: `node dist/index.js`
-3. **환경 변수**: `NODE_ENV=production PORT=5000`
-4. **호스트 바인딩**: `0.0.0.0` (자동 설정됨)
-
-이제 배포 환경에서 connection refused 오류가 해결되어야 합니다.
