--- DEPLOYMENT_ERRORS_FIXED.md
@@ -1,71 +1,0 @@
-# 배포 오류 완전 해결
-
-## 해결된 문제들
-
-### 1. 포트 연결 거부 오류 해결 ✅
-**문제**: `dial tcp 127.0.0.1:5000: connect: connection refused`
-**원인**: 서버가 잘못된 포트/주소로 바인딩
-**해결**: 
-- PORT 환경변수 올바른 사용 확인
-- 0.0.0.0 바인딩으로 접근성 보장
-- 테스트 완료: PORT=3000에서 정상 작동
-
-### 2. 모듈 경로 오류 해결 ✅  
-**문제**: `Cannot find module '/home/runner/workspace/dist/index.js'`
-**원인**: package.json의 start 스크립트 경로 오류
-**해결**:
-- dist/package.json에서 start 스크립트를 `node index.js`로 수정
-- 빌드 프로세스에서 올바른 경로 구조 생성
-
-## 배포 명령어
-
-### Replit 배포용 빌드
-```bash
-node deploy-fix-complete.js
-```
-
-### 수동 빌드 (백업용)
-```bash
-node build-simple.js
-```
-
-## 배포 검증 결과
-
-### 성공적으로 확인된 사항
-- ✅ dist/index.js 생성 (81.2KB)
-- ✅ 올바른 production package.json 생성
-- ✅ PORT 환경변수 처리 정상
-- ✅ 0.0.0.0:3000 바인딩 성공
-- ✅ 프로덕션 모드 서버 시작 확인
-- ✅ 모든 필수 파일 복사 완료
-
-### 배포 구조
-```
-dist/
-├── index.js          # 번들된 서버 (81.2KB)
-├── package.json      # 프로덕션 설정
-├── shared/           # DB 스키마
-└── public/           # 정적 파일
-```
-
-## Replit 배포 설정
-
-### 권장 .replit 설정
-```ini
-[deployment]
-deploymentTarget = "autoscale"
-build = ["node", "deploy-fix-complete.js"]
-run = ["npm", "run", "start"]
-```
-
-### 환경변수 설정 필요
-```
-DATABASE_URL=your_database_url
-SESSION_SECRET=your_secure_secret
-PORT=auto  # Replit이 자동 설정
-```
-
-## 문제 해결 완료
-
-모든 연결 거부 오류와 모듈 경로 문제가 해결되었습니다. 
-배포는 `node deploy-fix-complete.js` 명령어로 실행하면 됩니다.
