--- GIT_COMMANDS.md
@@ -1,61 +1,0 @@
-# Git Commands for New Branch
-
-## 현재 상황
-배포 오류 수정과 Vercel 설정이 완료된 상태입니다.
-
-## 새 브랜치 생성 및 푸시 명령어
-
-### 1. Git 상태 확인
-```bash
-git status
-```
-
-### 2. 변경사항 스테이징
-```bash
-git add .
-```
-
-### 3. 현재 변경사항 커밋
-```bash
-git commit -m "Fix deployment errors and add Vercel configuration
-
-- Fixed port connection refused errors with proper PORT handling
-- Resolved module path issues in production build
-- Added Vercel deployment configuration (vercel.json, api/index.ts)
-- Created deployment scripts (deploy-fix-complete.js, build-simple.js)
-- Added comprehensive deployment documentation
-- Verified production server starts correctly with 0.0.0.0 binding"
-```
-
-### 4. 새 브랜치 생성 및 전환
-```bash
-git checkout -b deployment-fixes
-```
-
-### 5. 새 브랜치를 원격 저장소에 푸시
-```bash
-git push -u origin deployment-fixes
-```
-
-## 또는 한 번에 실행
-```bash
-git add . && \
-git commit -m "Fix deployment errors and add Vercel configuration" && \
-git checkout -b deployment-fixes && \
-git push -u origin deployment-fixes
-```
-
-## 생성된 주요 파일들
-- ✅ vercel.json (Vercel 배포 설정)
-- ✅ api/index.ts (서버리스 함수)
-- ✅ deploy-fix-complete.js (배포 빌드 스크립트)
-- ✅ DEPLOYMENT_ERRORS_FIXED.md (해결 문서)
-- ✅ VERCEL_DEPLOYMENT.md (Vercel 가이드)
-- ✅ .env.example (환경변수 템플릿)
-
-## 브랜치 목적
-이 브랜치는 다음을 포함합니다:
-- 모든 배포 오류 수정
-- Vercel 배포 준비 완료
-- Replit 배포 최적화
-- 포트 및 모듈 경로 문제 해결
