--- REPLIT_DEPLOYMENT_FINAL_SOLUTION.md
@@ -1,46 +1,0 @@
-# Replit Autoscale 배포 문제 최종 해결
-
-## 문제 진단
-- **Connection Refused**: 프록시가 127.0.0.1:5000에 연결 시도하지만 서버는 다른 포트에서 실행
-- **MODULE_NOT_FOUND**: `/home/runner/workspace/dist/index.js` 파일을 찾을 수 없음
-- **Crash Loop**: 서버 시작 실패로 인한 반복적 재시작
-- **Build Process**: package.json의 build 명령어가 Replit autoscale과 호환되지 않음
-
-## 최종 해결책
-
-### 1. 포트 구성 완전 수정 ✅
-- 서버가 0.0.0.0:8080에 바인딩하도록 강제 설정
-- PORT=8080 환경변수를 빌드 시점에 주입
-- Replit autoscale 프록시 요구사항 충족
-
-### 2. 빌드 프로세스 재구성 ✅
-- `build.js` 스크립트로 package.json build 명령어 대체
-- 46.3kb 최적화된 번들 생성 (이전 81.3kb에서 43% 감소)
-- minify 적용으로 성능 향상
-
-### 3. 의존성 최적화 ✅
-- production package.json에 필수 런타임 의존성만 포함
-- external packages 처리로 번들 크기 최소화
-- Node.js 18+ 엔진 요구사항 명시
-
-### 4. 배포 구조 완성 ✅
-- `/home/runner/workspace/dist/index.js` 경로에 정확한 파일 생성
-- `npm run start` 명령어가 PORT=8080으로 서버 시작
-- 헬스 체크 엔드포인트 `/health` 포함
-
-## 검증 결과
-✅ dist/index.js 파일이 올바른 위치에 생성됨
-✅ 서버가 0.0.0.0:8080에서 정상 시작
-✅ MODULE_NOT_FOUND 오류 완전 해결
-✅ Connection refused 오류 해결
-✅ Crash loop 방지 구현
-✅ 46.3kb 최적화된 번들로 배포 속도 향상
-
-## 배포 프로세스
-Replit autoscale 배포 시:
-1. `npm run build` → `node build.js` 실행
-2. `dist/` 디렉토리에 최적화된 번들 생성
-3. `npm run start` → `PORT=8080 NODE_ENV=production node index.js` 실행
-4. 서버가 0.0.0.0:8080에서 시작되어 프록시 연결 처리
-
-이제 "dial tcp 127.0.0.1:5000: connect: connection refused"와 "Cannot find module" 오류 없이 정상 배포됩니다.
