--- REPLIT_PORT_5000_FINAL_FIX.md
@@ -1,50 +1,0 @@
-# Replit Autoscale Port 5000 문제 최종 해결
-
-## 문제 분석
-1. `.replit` 파일에서 port 5000이 external port 80에 매핑되어 있어 Replit이 5000번 포트를 기대
-2. 애플리케이션은 8080에서 실행되지만 배포 시스템이 5000번 포트에서 연결 시도
-3. 빌드 프로세스가 올바른 dist/index.js 파일을 생성하지만 포트 충돌 발생
-
-## 해결된 사항 ✅
-
-### 1. 빌드 프로세스 완전 수정
-- `build.js` 스크립트가 46.3kb 최적화된 번들 생성
-- 포트 8080이 빌드 시점에 하드코딩됨
-- `dist/index.js` 파일이 올바른 위치에 생성됨
-
-### 2. package.json 배포 구성
-- `npm run start` 명령어가 `PORT=8080 NODE_ENV=production node index.js` 실행
-- 필수 런타임 의존성만 포함
-- Node.js 18+ 엔진 요구사항 설정
-
-### 3. 서버 코드 포트 강제 설정
-- 모든 포트 참조를 8080으로 하드코딩
-- 0.0.0.0 호스트 바인딩으로 외부 접근 허용
-- 프로덕션 환경에서 포트 변경 방지
-
-## 배포 프로세스 검증 ✅
-
-1. **빌드 단계**: `npm run build` → `node build.js`
-   - ✅ dist/index.js 생성 (46.3kb)
-   - ✅ 포트 8080 하드코딩 적용
-   - ✅ 프로덕션 package.json 생성
-
-2. **시작 단계**: `npm run start` 
-   - ✅ PORT=8080 환경변수 설정
-   - ✅ 서버가 0.0.0.0:8080에서 시작
-   - ✅ MODULE_NOT_FOUND 오류 해결
-
-## 남은 이슈
-
-Replit autoscale이 여전히 포트 5000을 기대하는 이유:
-- `.replit` 파일의 포트 매핑 구성
-- 배포 시스템이 기존 포트 설정을 참조
-
-## 최종 권장사항
-
-배포 시 다음 사항을 확인:
-1. 애플리케이션이 8080에서 정상 시작되는지 확인
-2. 배포 로그에서 "Server successfully started on 0.0.0.0:8080" 메시지 확인
-3. connection refused 오류가 발생하면 포트 매핑 재구성 필요
-
-현재 빌드는 포트 8080에서 정상 작동하도록 완전히 구성되었습니다.
