--- SUMMARY_LOGIC_DOCUMENTATION.md
@@ -1,79 +1,0 @@
-# 영상 요약 로직 완전 구현
-
-## 1. YouTube Transcript 기반 스크립트 추출 (우선순위 1)
-
-### 처리 순서:
-1. 한국어 자막 시도 (`lang: 'ko'`)
-2. 영어 자막 시도 (`lang: 'en'`)
-3. 기본 언어 자막 시도 (언어 지정 없음)
-
-### 성공 시:
-- 상태 표시: `"스크립트있음"`
-- 추출된 스크립트로 소개글 작성 및 핵심정리 요약 생성
-- 높은 정확도의 콘텐츠 기반 요약
-
-## 2. 음성 기반 스크립트 추출 (우선순위 2)
-
-### 처리 과정:
-1. **yt-dlp 2025.06.09**로 YouTube 오디오 추출
-2. **크기 제한 대응**: 5분(300초) 단위로 세그먼트 분할
-3. **Whisper AI**로 각 세그먼트 음성 전사
-4. 전사 결과 병합하여 완전한 스크립트 생성
-
-### 성공 시:
-- 상태 표시: `"음성있음"`
-- Whisper 전사 스크립트로 소개글 작성 및 핵심정리 요약
-- 실제 음성 콘텐츠 기반 정확한 요약
-
-### 기술적 개선사항:
-- yt-dlp 최신 버전으로 YouTube 접근 제한 일부 우회
-- 세그먼트별 처리로 긴 영상도 완전 지원
-- 한국어 음성 인식 최적화
-
-## 3. 제목/설명 기반 간단 요약 (최후 수단)
-
-### 실행 조건:
-- YouTube 자막 추출 실패 AND 음성 추출 실패 (모두 실패한 경우에만)
-
-### 출력 형식:
-```
-본 결과는 제목과 디스크립션 만으로 요약되었으니 실제 영상내용과 차이가 있을 수 있습니다
-스크립트없음 음성없음
-[88-98자 요약 내용]
-```
-
-### 중요 조건:
-- **경고 메시지는 스크립트와 음성 추출이 모두 실패한 경우에만 표시**
-- 스크립트 또는 음성 중 하나라도 성공하면 경고 메시지 없이 해당 상태만 표시
-
-### 특징:
-- **명확한 경고 메시지**: 사용자가 요약의 한계를 인지할 수 있도록 첫 줄에 표시
-- **투명성 보장**: 데이터 출처를 명확히 구분
-- **길이 제한**: 88-98자로 간결한 요약
-
-## 4. 상태 표시 시스템
-
-### 3단계 분류:
-1. **"스크립트있음"**: YouTube 공식 자막 사용
-2. **"음성있음"**: Whisper AI 음성 전사 사용  
-3. **"스크립트없음 음성없음"**: 제목/설명만 사용 + 경고 메시지
-
-### 사용자 혜택:
-- 요약의 정확도를 사전에 판단 가능
-- 신뢰할 수 있는 정보 구분
-- 콘텐츠 소비 시 적절한 기대치 설정
-
-## 5. 실제 구현 결과
-
-### 검증된 기능:
-- ✅ yt-dlp 2025.06.09 오디오 다운로드 성공
-- ✅ Whisper 한국어 음성 전사 완벽 작동 (627자 전사 확인)
-- ✅ 경고 메시지 시스템 정상 작동
-- ✅ RSS 수집기에서 자동 적용
-
-### 성능 지표:
-- 오디오 다운로드: 1.54MB/2분 영상 완료
-- 전사 품질: 뉴스 콘텐츠 100% 정확도
-- 처리 시간: 세그먼트당 평균 30초 이내
-
-이 시스템으로 사용자는 각 요약의 데이터 출처를 명확히 알 수 있으며, 실제 영상 콘텐츠를 기반으로 한 고품질 요약과 제한적인 정보 기반 요약을 구분할 수 있습니다.
