--- deploy.js
@@ -1,61 +1,0 @@
-#!/usr/bin/env node
-// Deployment script that replaces the broken npm run build
-// This script ensures dist/index.js is created correctly for deployment
-import { execSync } from 'child_process';
-import fs from 'fs';
-
-console.log('Starting deployment build process...');
-
-try {
-  // Clean dist directory
-  if (fs.existsSync('dist')) {
-    fs.rmSync('dist', { recursive: true, force: true });
-  }
-  fs.mkdirSync('dist', { recursive: true });
-
-  // Build server bundle using esbuild with correct parameters
-  console.log('Building server bundle...');
-  execSync('esbuild server/index.ts --platform=node --packages=external --bundle --format=esm --outfile=dist/index.js --alias:@shared=./shared', {
-    stdio: 'inherit'
-  });
-
-  // Create production package.json
-  const packageJson = JSON.parse(fs.readFileSync('package.json', 'utf8'));
-  const prodPackageJson = {
-    name: packageJson.name,
-    version: packageJson.version,
-    type: packageJson.type,
-    scripts: {
-      start: "node index.js"
-    },
-    dependencies: packageJson.dependencies,
-    optionalDependencies: packageJson.optionalDependencies || {}
-  };
-
-  fs.writeFileSync('dist/package.json', JSON.stringify(prodPackageJson, null, 2));
-
-  // Copy required files
-  if (fs.existsSync('shared')) {
-    fs.cpSync('shared', 'dist/shared', { recursive: true });
-  }
-  
-  fs.mkdirSync('dist/public', { recursive: true });
-  
-  if (fs.existsSync('client')) {
-    fs.cpSync('client', 'dist/client', { recursive: true });
-  }
-
-  // Critical verification
-  if (!fs.existsSync('dist/index.js')) {
-    console.error('DEPLOYMENT FAILURE: dist/index.js was not created');
-    process.exit(1);
-  }
-
-  const stats = fs.statSync('dist/index.js');
-  console.log(`Deployment build successful - dist/index.js created (${(stats.size / 1024).toFixed(1)}kb)`);
-  console.log('Ready for deployment with "node dist/index.js"');
-
-} catch (error) {
-  console.error('Deployment build failed:', error.message);
-  process.exit(1);
-}
